Update to ns-3-dev

+++ src/wifi/model/yans-error-rate-model.h
@@ -21,8 +21,6 @@
 #ifndef YANS_ERROR_RATE_MODEL_H
 #define YANS_ERROR_RATE_MODEL_H
 
-#include <stdint.h>
-#include "wifi-mode.h"
 #include "error-rate-model.h"
 #include "dsss-error-rate-model.h"
 
@@ -56,6 +54,10 @@ namespace ns3 {
 class YansErrorRateModel : public ErrorRateModel
 {
 public:
+  /**
+   * \brief Get the type ID.
+   * \return the object TypeId
+   */
   static TypeId GetTypeId (void);
 
   YansErrorRateModel ();
@@ -81,7 +83,7 @@ private:
    *
    * \return BER of BPSK at the given SNR
    */
-  double GetBpskBer (double snr, uint32_t signalSpread, uint32_t phyRate) const;
+  double GetBpskBer (double snr, uint32_t signalSpread, uint64_t phyRate) const;
   /**
    * Return BER of QAM-m with the given parameters.
    *
@@ -92,7 +94,7 @@ private:
    *
    * \return BER of BPSK at the given SNR
    */
-  double GetQamBer (double snr, unsigned int m, uint32_t signalSpread, uint32_t phyRate) const;
+  double GetQamBer (double snr, unsigned int m, uint32_t signalSpread, uint64_t phyRate) const;
   /**
    * Return k!
    *
@@ -143,7 +145,7 @@ private:
    * \return double
    */
   double GetFecBpskBer (double snr, double nbits,
-                        uint32_t signalSpread, uint32_t phyRate,
+                        uint32_t signalSpread, uint64_t phyRate,
                         uint32_t dFree, uint32_t adFree) const;
   /**
    * \param snr SNR ratio (not dB)
@@ -159,7 +161,7 @@ private:
    */
   double GetFecQamBer (double snr, uint32_t nbits,
                        uint32_t signalSpread,
-                       uint32_t phyRate,
+                       uint64_t phyRate,
                        uint32_t m, uint32_t dfree,
                        uint32_t adFree, uint32_t adFreePlusOne) const;
 };
